+++ core/domain/src/commonMain/kotlin/com/softartdev/notedelight/usecase/settings/AppVersionUseCase.kt
@@ -1,0 +1,5 @@
+package com.softartdev.notedelight.usecase.settings
+
+expect class AppVersionUseCase {
+    operator fun invoke(): String?
+}