+++ core/domain/src/wasmJsMain/kotlin/com/softartdev/notedelight/usecase/settings/AppVersionUseCase.wasmJs.kt
@@ -1,0 +1,6 @@
+package com.softartdev.notedelight.usecase.settings
+
+actual class AppVersionUseCase {
+
+    actual operator fun invoke(): String? = null
+}