+++ thirdparty/androidx/paging/compose/src/wasmJsMain/kotlin/PagingPlaceholders.wasmJs.kt
@@ -1,3 +1,0 @@
-package androidx.paging.compose
-
-internal actual fun getPagingPlaceholderKey(index: Int): Any = index