+++ ui/test/src/commonMain/kotlin/com/softartdev/notedelight/CommonUiTests.kt
@@ -119,6 +119,12 @@ abstract class CommonUiTests : AbstractUITests() {
         super.localeTest()
     }
 
+    override fun backupFeatureTest(): TestResult = runComposeUiTest {
+        _composeUiTest = this
+        beforeTest()
+        super.backupFeatureTest()
+    }
+
     override fun pressBack() {
         composeUiTest.onNodeWithContentDescription(label = Icons.AutoMirrored.Filled.ArrowBack.name)
             .assertIsDisplayed()