+++ ui/test/src/commonMain/kotlin/com/softartdev/notedelight/ui/screen/SettingsTestScreen.kt
@@ -10,18 +10,18 @@ import com.softartdev.notedelight.util.EXPORT_DATABASE_BUTTON_TAG
 import com.softartdev.notedelight.util.IMPORT_DATABASE_BUTTON_TAG
 import com.softartdev.notedelight.util.LANGUAGE_BUTTON_TAG
 import com.softartdev.notedelight.util.SET_PASSWORD_BUTTON_TAG
+import com.softartdev.notedelight.util.SETTINGS_CATEGORY_APPEARANCE_TAG
 import com.softartdev.notedelight.util.SETTINGS_CATEGORY_BACKUP_TAG
 import com.softartdev.notedelight.util.SETTINGS_CATEGORY_INFO_TAG
 import com.softartdev.notedelight.util.SETTINGS_CATEGORY_SECURITY_TAG
-import com.softartdev.notedelight.util.SETTINGS_CATEGORY_THEME_TAG
 import kotlin.jvm.JvmInline
 
 @JvmInline
 value class SettingsTestScreen(val nodeProvider: SemanticsNodeInteractionsProvider) {
 
-    val themeCategorySNI: SemanticsNodeInteraction
+    val appearanceCategorySNI: SemanticsNodeInteraction
         get() = nodeProvider
-            .onNodeWithTag(SETTINGS_CATEGORY_THEME_TAG)
+            .onNodeWithTag(SETTINGS_CATEGORY_APPEARANCE_TAG)
             .assertIsDisplayed()
 
     val securityCategorySNI: SemanticsNodeInteraction